fix: se realizo restructuracion y renombramiento de variables, se realiza validacion de redireccionamiento
