+++ etc/adminhtml/system.xml
@@ -5,38 +5,37 @@
  * See COPYING.txt for license details.
  */
 -->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../Config/etc/system_file.xsd">
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Config:etc/system_file.xsd">
     <system>
         <section id="payment">
-            <group id="paycoagregador" translate="label" type="text" sortOrder="100" showInDefault="1" showInWebsite="1" showInStore="1">
-
-               <label>Paycoagregador v2.0.0</label>
-                <field id="active" translate="label" type="select" sortOrder="1" showInDefault="1" showInWebsite="1" showInStore="0">
-                    <label>Habilitar/Deshabilitar</label>
+            <group id="epaycoagregador" translate="label" type="text" sortOrder="100" showInDefault="1" showInWebsite="1" showInStore="1">
+               <label>ePayco agregador</label>
+                <field id="active" translate="label" type="select" sortOrder="1" showInDefault="1" showInWebsite="1" showInStore="0" canRestore="1">
+                    <label>Habilitar</label>
                     <comment>Habilitar o deshabilitar ePayco Checkout</comment>
                     <source_model>Magento\Config\Model\Config\Source\Yesno</source_model>
                 </field>
-                <field id="paycoagregador_title" translate="label" type="text" sortOrder="2" showInDefault="1" showInWebsite="1" showInStore="0">
+                <field id="payco_title" translate="label" type="text" sortOrder="2" showInDefault="1" showInWebsite="1" showInStore="0">
                     <label>Título</label>
                     <comment>Corresponde al titulo que el usuario ve durante el checkout.</comment>
                 </field>
-                <field id="paycoagregador_merchant" translate="label" type="text" sortOrder="4" showInDefault="1" showInWebsite="1" showInStore="0">
+                <field id="payco_merchant" translate="label" type="text" sortOrder="4" showInDefault="1" showInWebsite="1" showInStore="0">
                     <label>P_CUST_ID_CLIENTE</label>
                     <comment>ID de cliente que lo identifica en ePayco. Lo puede encontrar en su panel de clientes en la opción configuración.</comment>
                 </field>
-                <field id="paycoagregador_key" translate="label" type="text" sortOrder="5" showInDefault="1" showInWebsite="1" showInStore="0">
+                <field id="payco_key" translate="label" type="text" sortOrder="5" showInDefault="1" showInWebsite="1" showInStore="0">
                     <label>P_KEY</label>
                     <comment>Llave para firmar la información enviada y recibida de ePayco. Lo puede encontrar en su panel de clientes en la opción configuración.</comment>
                 </field>
-                <field id="paycoagregador_public_key" translate="label" type="text" sortOrder="6" showInDefault="1" showInWebsite="1" showInStore="0">
+                <field id="payco_public_key" translate="label" type="text" sortOrder="6" showInDefault="1" showInWebsite="1" showInStore="0">
                     <label>Public_KEY</label>
                     <comment>LLave para autenticar y consumir los servicios de ePayco, Proporcionado en su panel de clientes en la opción configuración.</comment>
                 </field>
-                <field id="paycoagregador_callback" translate="label" type="text" sortOrder="7" showInDefault="1" showInWebsite="1" showInStore="0">
+                <field id="payco_callback" translate="label" type="text" sortOrder="7" showInDefault="1" showInWebsite="1" showInStore="0">
                     <label>URL De Respuesta</label>
                     <comment>Url de respuesta para confirmar los pagos desde ePayco.</comment>
                 </field>
-                <field id="paycoagregador_test" translate="label" type="select" sortOrder="9" showInDefault="1" showInWebsite="1" showInStore="0">
+                <field id="payco_test" translate="label" type="select" sortOrder="9" showInDefault="1" showInWebsite="1" showInStore="0">
                     <label>Modo Pruebas</label>
                     <comment>Habilitar el envio de pago en modo pruebas</comment>
                     <source_model>Magento\Config\Model\Config\Source\Yesno</source_model>
@@ -46,11 +45,15 @@
                     <comment>(Onpage Checkout, el usuario a pagar permanece en el sitio) o (Standart Checkout, el usuario a pagar se redirige a la puerta de enlace de ePayco)</comment>
                     <source_model>Pago\Paycoagregador\Model\Config\Source\Helper\Vertical</source_model>
                 </field>
-                <field id="paycoagregador_sort_order" translate="label" type="text" sortOrder="11" showInDefault="1" showInWebsite="1" showInStore="0">
-                    <label>Orden</label>
+                <field id="order_status" translate="label" type="select" sortOrder="20" showInDefault="1" showInWebsite="1" showInStore="0" canRestore="1">
+                    <label>New Order Status</label>
+                    <source_model>Magento\Sales\Model\Config\Source\Order\Status\NewStatus</source_model>
+                </field>
+                <field id="sort_order" translate="label" type="text" sortOrder="100" showInDefault="1" showInWebsite="1" showInStore="0">
+                    <label>Sort Order</label>
+                    <frontend_class>validate-number</frontend_class>
                 </field>
             </group>
-            <!-- payment-group -->
         </section>
     </system>
 </config>
