+++ etc/frontend/di.xml
@@ -1,10 +1,15 @@
 <?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-
     <type name="Magento\Checkout\Model\CompositeConfigProvider">
         <arguments>
             <argument name="configProviders" xsi:type="array">
-                <item name="custom_paycoagregador_config_provider" xsi:type="object">Pago\Paycoagregador\Model\PaycoagregadorConfigProvider</item>
+                <item name="online_payment_epaycoagregador_config_provider" xsi:type="object">Pago\Paycoagregador\Model\CustomConfigProvider</item>
             </argument>
         </arguments>
     </type>
