+++ etc/frontend/routes.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" ?>
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
     <router id="standard">
-        <route frontName="confirmation" id="confirmation">
+        <route frontName="confirmationAgregador" id="confirmationAgregador">
             <module name="Pago_Paycoagregador"/>
         </route>
-        <route frontName="response" id="response">
+        <route frontName="responseAgregador" id="responseAgregador">
             <module name="Pago_Paycoagregador"/>
         </route>
     </router>
