+++ etc/module.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0"?>
 <!--
 /**
-* @copyright Copyright (c) 2015 X.commerce, Inc. (http://www.magentocommerce.com)
-*/
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
 -->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../lib/internal/Magento/Framework/Module/etc/module.xsd">
-  <module name="Pago_Paycoagregador" setup_version="2.0.0">
-  	 		<module name="Magento_Sales" />
-            <module name="Magento_Checkout" />
-            <module name="Magento_Payment" />
-            <module name="Magento_Directory" />
-			<module name="Magento_Config" />
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
+  <module name="Pago_Paycoagregador" setup_version="2.0.2">
+    <sequence>
+      <module name="Magento_Store"/>
+      <module name="Magento_Catalog"/>
+    </sequence>
   </module>
 </config>
