+++ etc/payment.xml
@@ -6,16 +6,15 @@
  */
 -->
 <payment xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:noNamespaceSchemaLocation="../../../Magento/Payment/etc/payment.xsd">
+         xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Payment:etc/payment.xsd">
     <groups>
-        <group id="online">
-            <label>Online Payment Methods</label>
+        <group id="epaycoagregador_payment">
+            <label>Online Payment Methods epayco</label>
         </group>
     </groups>
     <methods>
-        <method name="Paycoagregador">
+        <method name="epaycoagregador">
             <allow_multiple_address>1</allow_multiple_address>
         </method>
-
     </methods>
 </payment>
