+++ view/frontend/layout/confirmationAgregador_epaycoagregador_index.xml
@@ -2,7 +2,7 @@
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"  layout="1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
     <body>
         <referenceContainer name="content">
-            <block class="Pago\Paycoagregador\Block\Paycoagregador\Action" name="paycoagregador_payment" template="Pago_Paycoagregador::payment.phtml"/>
+            <block class="Pago\Paycoagregador\Block\Epaycoagregador\Action" name="epaycoagregador_confirmationAgregador" template="Pago_Paycoagregador::confirmation.phtml"/>
         </referenceContainer>
     </body>
 </page>
