--- view/frontend/layout/response_paymentagregador_index.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" ?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"  layout="1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceContainer name="content">
-            <block class="Pago\Paycoagregador\Block\Paycoagregador\Action" name="paycoagregador_paymentagregador" template="Pago_Paycoagregador::response.phtml"/>
-        </referenceContainer>
-    </body>
-</page>
-
-
-
-
-
