+++ view/frontend/templates/confirmation.phtml
@@ -1,10 +1,10 @@
 <?php
 	/**
-	 * @var \Pago\Paycoagregador\Block\Paycoagregador\Index $block
+	 * @var \PagoEpayco\Payco\Block\Epayco\Index $block
 	 *
 	 */
-
-
+	
+	
 	var_dump('hola');
 	die();
  echo $block->displayConfirmation();?>
@@ -12,4 +12,4 @@
 <div>
 	juan david
 </div>
-
+
