--- view/frontend/templates/payment.phtml
@@ -1,15 +1,0 @@
-<?php
-/**
- * @var \Pago\Paycoagregador\Block\Paycoagregador\Index $block
- *
- */
-
-
-var_dump('hola');
-die();
-echo $block->displayConfirmation();?>
-
-<div>
-    juan david
-</div>
-
