+++ view/frontend/web/template/payment/epaycoagregador.html
@@ -1,9 +1,3 @@
-<!--
-/**
- * Copyright © 2015 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
 <div class="payment-method" data-bind="css: {'_active': (getCode() == isChecked())}">
     <div class="payment-method-title field choice">
         <input type="radio"
@@ -23,11 +17,15 @@
         </div>
         <div class="checkout-agreements-block">
             <!-- ko foreach: $parent.getRegion('before-place-order') -->
-                <!-- ko template: getTemplate() --><!-- /ko -->
+            <!-- ko template: getTemplate() --><!-- /ko -->
             <!--/ko-->
         </div>
         <div class="actions-toolbar">
             <div class="primary">
+
+                <!--<input type="image" id="imagen" src="https://369969691f476073508a-60bf0867add971908d4f26a64519c2aa.ssl.cf5.rackcdn.com/btns/btn4.png" data-bind="
+    click: renderCheckout
+   "/> -->
                 <button class="action primary checkout"
                         type="submit"
                         data-bind="
@@ -39,7 +37,12 @@
                         disabled>
                     <span data-bind="i18n: 'Place Order'"></span>
                 </button>
+                <!-- <button class="action primary checkout"
+				 type="button" onclick="renderCheckout()">
+				 <span data-bind="i18n: 'Place Order'"></span>
+			   </button>-->
             </div>
         </div>
     </div>
+
 </div>
